Handle fractional late days.
refs: MBL-19082
affects: Teacher
release note: Added support for fractional late days number to SpeedGrader.

test plan:
- Change grade penalty to be measure in hours instead of days in web gradebook.
- Open web SpeedGrader and make the submission late by 12 hours.
- Open teacher app and check the same submission.
- Late days should show 0.5.
- Edit late days to 0.25 in the app.
- Go to web, refresh SpeedGrader.
- Late hours shold show 6 hours.
- Late penalty score should also match between app and web.

+++ Teacher/Teacher/SpeedGrader/Grading/GradeStatus/Model/GradeStatusInteractorPreview.swift
@@ -61,9 +61,9 @@ class GradeStatusInteractorPreview: GradeStatusInteractor {
         .none
     }
 
-    func observeGradeStatusChanges(submissionId: String, attempt: Int) -> AnyPublisher<(GradeStatus, daysLate: Int, dueDate: Date?), Never> {
+    func observeGradeStatusChanges(submissionId: String, attempt: Int) -> AnyPublisher<(GradeStatus, daysLate: Double, dueDate: Date?), Never> {
         if let status = previewStatus, let daysLate = previewDaysLate {
-            return Just((status, daysLate, previewDueDate)).eraseToAnyPublisher()
+            return Just((status, Double(daysLate), previewDueDate)).eraseToAnyPublisher()
         } else {
             return Empty().eraseToAnyPublisher()
         }
@@ -72,7 +72,7 @@ class GradeStatusInteractorPreview: GradeStatusInteractor {
     func updateLateDays(
         submissionId: String,
         userId: String,
-        daysLate: Int
+        daysLate: Double
     ) -> AnyPublisher<Void, Error> {
         Just(())
             .delay(for: .seconds(2), scheduler: RunLoop.main)
